+++ hubdle-gradle-plugin/testFunctional/resources/kotlin/jvm/features/power-assert/test/kotlin/com/javiersc/kotlin/jvm/power/assert/FooTest.kt
@@ -1,0 +1,14 @@
+package com.javiersc.kotlin.jvm.power.assert
+
+import kotlin.test.assertTrue
+import kotlin.test.Test
+
+class FooTest {
+
+    @Test
+    fun test() {
+        val foo = "FOO"
+        val bar = "BAR"
+        assertTrue(foo == bar)
+    }
+}